Creacion VistaJuego, PanelJuego, ControladorPanelJuego, Modelo PanelJuego con patron Observer
