+++ src/main/java/View/VistaJuego.java
@@ -1,15 +1,13 @@
 package View;
 
-import Model.GamePanel;
-
 import javax.swing.JFrame;
 
-public class Window extends JFrame {
+public class VistaJuego extends JFrame {
 
-    public Window() {
+    public VistaJuego(PanelJuego panelJuego) {
         setTitle("Bubble Bobble");
         setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        setContentPane(new GamePanel());
+        setContentPane(panelJuego);
         pack();
         setLocationRelativeTo(null);
         setResizable(false);
