+++ src/main/java/util/Observer.java
@@ -1,0 +1,6 @@
+package util;
+
+public interface Observer {
+
+    void update();
+}
