+++ src/main/java/util/Subject.java
@@ -1,0 +1,8 @@
+package util;
+
+public interface Subject {
+
+    boolean registerObserver(Observer o);
+    boolean removeObserver(Observer o);
+    void notifyObservers();
+}
