+++ src/test/java/PanelJuegoTest.java
@@ -1,0 +1,27 @@
+import View.PanelJuego;
+import View.VistaJuego;
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+public class PanelJuegoTest {
+
+    /**
+     * TODO Implementar corerctamente los casos de prueba
+     */
+//    @Test
+//    public void startGame(){
+//        VistaJuego vistaJuego = new VistaJuego();
+//        PanelJuego panelJuego = (PanelJuego) vistaJuego.getContentPane();
+//        assertTrue(panelJuego.running);
+//    }
+//
+//    @Test
+//    public void stopGame() {
+//        VistaJuego vistaJuego = new VistaJuego();
+//        PanelJuego panelJuego = (PanelJuego) vistaJuego.getContentPane();
+//        panelJuego.stopGame();
+//        assertFalse(panelJuego.running);
+//    }
+
+}
